remove unnecessary and buggy type checks in asyncio
Within asyncio, several `write`, `send` and `sendto` methods
contained a check whether the `data` parameter is one of the
*right* types, complaining that the data should be bytes-like.

This error message is very misleading if one hands over a byte-like
object that happens not to be one of the *right* types.

Usually, these kinds of tests are not necessary at all, as the
code the data is handed to will complain itself if the type is wrong.
But in the case of the mentioned methods, the data may be put into
a buffer if the data cannot be sent immediately. Any type errors
will only be raised much later, when the buffer is finally sent.

Interestingly, the test is incorrect as well: any memoryview is
considered *right*, although it may be passed to functions that
cannot deal with non-contiguous memoryviews, in which case the all
the problems that the test should protect from reappear.

There are several options one can go forward:

* one could update the documentation to reflect the fact that not
  all bytes-like objects can be passed, but only special ones. This
  would be unfortunate as the underlying code actually can deal with
  all bytes-like data.
* actually test whether the data is bytes-like. This is unfortunately
  not easy. The correct test would be to check whether the data can
  be parsed as by PyArg_Parse with a y* format. I am not aware of
  a simple test like this.
* Remove the test. In this case one should assure that only bytes-like
  data will be put into the buffers if the data cannot be sent
  immediately.

For simplicity I opted for the last option. One should note another
problem here: if we accept objects like memoryview or bytearray
(which we currently do), the user may modify the data after-the-fact,
which will lead to weird, unexpected behavior. This could be mitigated
by always copying the data. This is done in some of the modified
methods, but not all, most likely for performance reasons. I think
it would be beneficial to deal with this problem in a systematic way,
but this is beyond the scope of this patch.