GH-139174: Prepare `pathlib.Path.info` for new methods
Merge `_WindowsPathInfo` and `_PosixPathInfo` classes into a new
`StatResultInfo` class. On Windows, this means relying on `os.stat()`
rather than `os.path.isfile()` and friends, which is a little slower. But
there's value in making the code easier to maintain, and we're going to
need the stat result for implementing `size()`, `mode()` etc.