GH-139174: Add `pathlib.Path.info.stat()`
Add `stat()` method to `pathlib.Path.info` that returns a (possibly cached)
`os.stat_result` object. We don't add it to `pathlib.types.PathInfo`
because it's too specific to local filesystem paths. This requires a bit of
reworking of the docs to explain!

Rename `pathlib._Info` to `pathlib.Info` and document it, including the new
`stat()` method. Ensure it can't be instantiated by users. Move the
existing docs for `pathlib.types` to a new page.