GH-139282: Support recursive wildcard in `pathlib.PurePath.match()`
Add support for the recursive wildcard `**` in `pathlib.PurePath.match()`.
This is a behaviour change for anyone currently calling `match()` with `**`
in their pattern, but I reckon the vast majority of those users are
expecting recursive matching and haven't realised that it's not working (:

This brings `match()` behaviour closer to `full_match()`.