gh-139283: correctly handle `size` limit in `cursor.fetchmany()` (GH-139296)
Passing a negative or zero size to `cursor.fetchmany()` made it fetch all rows
instead of none.

While this could be considered a security vulnerability, it was decided to treat
this issue as a regular bug as passing a non-sanitized *size* value in the first
place is not recommended.
(cherry picked from commit bc172ee8307431caf4c89612e9e454081635191f)

Co-authored-by: Bénédikt Tran <10796600+picnixz@users.noreply.github.com>