Update tests for zip files with make_archive()
Improved readability and added comments to _create_files_symlinks() for making a test directory structure.

Expanded docstring for test_make_zipfile_symlink_behaviour to explain purpose of test and reasons why the test might be skipped in a Windows environment.

Updated test with try-finally for proper clean up of files after the test.