Use PyBytesWriter in pycore_blocks_output_buffer.h
Currently, the _BlocksOutputBuffer code creates a list of bytes
objects to handle the output data from compression libraries. This ends
up being slow due to the output buffer code needing to copy each bytes
element of the list into the final bytes object buffer at the end of
compression.

The new PyBytesWriter API introduced in PEP 782 is an ergonomic and fast
method of writing data into a buffer that will later turn into a bytes
object. Benchmarks show that using the PyBytesWriter API is 10-30%
faster for decompression across a variety of settings. The performance
gains are greatest when the decompressor is very performant, such as for
Zstandard. Otherwise the decompressor can bottleneck decompression and
the gains are more modest, but still sizable!