gh-84027: Add whl as an unpack format in shutil
Wheels are the most ubiquitous Python package format. Per [the wheel
format
specification](https://packaging.python.org/en/latest/specifications/binary-distribution-format/),
"A wheel is a ZIP-format archive with a specially formatted file name
and the .whl extension."

To enable unpacking of wheels for inspection or even installation, we
register the format in shutil using shutil._unpack_zipfile to do the
actual unpacking.