gh-141713: Prevent possible memory denial of service when reading
Make read() and similar methods that read a given number of bytes use
a progressively growing buffer instead of allocating the maximum size buffer
at once.  This helps prevent certain kind of memory denial of service issues
when the number of bytes to read within a specific protocol or format is
received from untrusted source.