gh-141786 Fix missing parent executor during trace
An executor's trace can point to another executor, forming a graph
of traces / executors. Sometimes it is possible while recording a
trace that the parent executor is freed / invalidated halfway. This
leads to the bug described in the issue gh-141786, which if left
unfixed could cause a memory leak.

This patch checks for the validity of the parent executor as well
as allowing JIT from the cold executor. While the cold executor is
not linked, it is the executor responsible for creating side traces
and we still want to JIT from it.