configure.ac: detect unusable termio operations
Some termio operations are not actually usable on some
architectures. For example, the TCGETA, TCSETA, TCSETAF and TCSETAW
are defined with a reference to "struct termio" on alpha, hppa and
sparc64, but "struct termio" is no longer defined since glibc 2.42,
causing a build failure.

Instead of using those operations as soon as they are defined, this
commit checks more carefully that they are actually usable. Such a
check cannot be done with the standard AC_CHECK_DECL() m4 macro, as
this macro only checks that the #define exists, not that it is
actually usable (i.e builds properly). Therefore, we introduce a
PY_CHECK_CONSTANT() m4 macro, that verifies that the constant is
actually usable.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>