Document new API pattern, add deprecation helpers.
The goal of this refactoring is to solve a non-linear complexity issue
that arises from the fact that the current API does at *least* one
substring slice per function call, and the depth of the call stack can
get fairly deep when dealing with a complex header.  (The complexity
isn't quadratic, since not every character in a header triggers a slice,
and python tries to optimize string manipulations, but it is
non-linear.)  The current parser is also inefficient both in memory use
and processing speed.

This rewrite, which I had been planning since writing the first version
of the parser way back when but hadn't gotten to until now, is quite
worthwhile even without the goal of reducing the non-linearity.

The old API for almost all the 'get' functions takes a 'value', parses
starting at its beginning, and returns an object representing the parsed
value plus the substring of value that wasn't parsed (that string copy
is the non-linearity).  The new API takes the full string being parsed
as 'value' plus a 'start' point for parsing, and instead of the trailing
substring returns a pointer to the character after the last one that was
parsed, or to after the end of the string (the len of value).

This commit updates the prefix comment for the parser to discuss the
target API rather than the old API.

It also defines deprecation helpers and tests for them.

We'll deprecate things by adding '_deprecated_' as prefix to the name,
and suggest replacements when possible by adding to the _REPLACED_NAMES
dictionary (the _replaced_with decorator does this for deprecated
functions).  This handles functions and constants that will be obsolete
by the end of the refactoring.

A function decorator, _deprecate_old_api, provides a way to support both
the old and the new API, issuing a deprecation message if the call is in
the form of the old API.  This only works on functions that conform to
the description above, though it can accommodate additional arguments
beyond the value or additional return values beyond the (value,
remainder) pair, as long as the second additional argument isn't an
integer.  Fortunately none of the existing functions has an integer as a
second argument.

There are few functions that currently return an empty value if given
empty or invalid input (value[start] isn't the beginning of what they
are supposed to parse).  They *should* be raising an error to stay
consistent with the API of the rest of the functions.  A specialized
decorator, _deprecate_old_api_and_lack_of_raise_on_invalid_input,
handles this specific case, allowing us to deprecate the buggy behavior
while having new style API calls get the exception.

*During* refactoring, the decorator '_deprecate' will be used to
indicate the intent to deprecate a function, while allowing existing
calls to it to continue to use the old name.  This will be replaced by
the '_deprecated_' prefix and/or the '_replaced_with` decorator by the
end of refactoring.

The refactoring procedure I settled on after several iterations is the
preceding refactoring of all of the tests to use the new test framework,
followed by refactoring the parser functions one by one, in order, since
their order in the file is bottom up (there was one exception, which
we've fixed.)

For the refactoring process, if the caller of a function has also been
refactored so that it is passing an index, the decorator will just call
the decorated function.  But if the caller hasn't been refactored yet,
the decorator converts the arguments into something the refactored
function will do the right thing with, and converts the results into
something the caller expects.  This allows for both a stepwise
refactoring process and for supporting a deprecation period in case
anyone has been brave enough to use this internal API in their own
programs.

The constraints on the refactoring are that the function needs to
continue to return an object representing an extracted substring for
whatever result it is supposed to produce, and that we maintain the
pattern that these are independent functions (not object methods) that
take the string we're operating on as input and return an updated
position as output.